# config/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "test",
        "version": "0.0.1",
        "title": "Spacebox API"
    },
    "host": "127.0.0.1:5002",
    "basePath": "/",
...
    "schemes": [
        "https",
        "http"
    ],
    "paths": {
        "/account/account_balance/{address}": {
            "get": {
                "tags": [
                    "account"
                ],
                "summary": "get account balance",
                "description": "",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "address",
                        "in": "path",
                        "description": "address",
                        "required": true,
                        "type": "string",
                        "format": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/ApiResponse"
                        }
                    }
                }
            }
        },
        "/account/validators/{address}": {
            "get": {
                "tags": [
                    "account"
                ],
                "summary": "get account validators",
                "description": "",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "address",
                        "in": "path",
                        "description": "address",
                        "required": true,
                        "type": "string",
                        "format": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/ApiResponse"
                        }
                    }
                }
            }
        },
        "/account/votes/{address}": {
            "get": {
                "tags": [
                    "account"
                ],
                "summary": "get account votes",
                "description": "",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "address",
                        "in": "path",
                        "description": "address",
                        "required": true,
                        "type": "string",
                        "format": "string"
                    },
                    {
                        "name": "proposal_id",
                        "in": "query",
                        "description": "proposal_id",
                        "required": false,
                        "type": "string",
                        "format": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/ApiResponse"
                        }
                    }
                }
            }
        },
        "/account/account_info/{address}": {
            "get": {
                "tags": [
                    "account"
                ],
                "summary": "get account info",
                "description": "",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "address",
                        "in": "path",
                        "description": "address",
                        "required": true,
                        "type": "string",
                        "format": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/ApiResponse"
                        }
                    }
                }
            }
        },
        "/gov/proposals": {
            "get": {
                "tags": [
                    "gov"
                ],
                "summary": "get proposals list",
                "description": "",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "limit",
                        "required": false,
                        "type": "string",
                        "format": "string"
                    },
                    {
                        "name": "offset",
                        "in": "query",
                        "description": "offset",
                        "required": false,
                        "type": "string",
                        "format": "string"
                    },
                    {
                        "name": "status__in",
                        "in": "query",
                        "description": "status__in",
                        "required": false,
                        "type": "string",
                        "format": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/ApiResponse"
                        }
                    }
                }
            }
        },
        "/gov/proposal/{id}": {
            "get": {
                "tags": [
                    "gov"
                ],
                "summary": "get proposal by id",
                "description": "",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id",
                        "required": false,
                        "type": "string",
                        "format": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/ApiResponse"
                        }
                    }
                }
            }
        },
        "/gov/votes": {
            "get": {
                "tags": [
                    "gov"
                ],
                "summary": "get votes for list of proposals",
                "description": "",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "limit",
                        "required": false,
                        "type": "string",
                        "format": "string"
                    },
                    {
                        "name": "offset",
                        "in": "query",
                        "description": "offset",
                        "required": false,
                        "type": "string",
                        "format": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/ApiResponse"
                        }
                    }
                }
            }
        },
        "/gov/votes/{id}": {
            "get": {
                "tags": [
                    "gov"
                ],
                "summary": "get for proposal by proposal id",
                "description": "",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id",
                        "required": true,
                        "type": "string",
                        "format": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/ApiResponse"
                        }
                    }
                }
            }
        },
        "/gov/votes/{id}/validators-info": {
            "get": {
                "tags": [
                    "gov"
                ],
                "summary": "get validators votes for proposal by proposal id",
                "description": "",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id",
                        "required": true,
                        "type": "string",
                        "format": "string"
                    },
                    {
                        "name": "validator_option",
                        "in": "query",
                        "description": "validator_option",
                        "required": false,
                        "type": "string",
                        "format": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/ApiResponse"
                        }
                    }
                }
            }
        },
        "/gov/votes/{id}/validators-info/{validator_address}": {
            "get": {
                "tags": [
                    "gov"
                ],
                "summary": "get validators votes for proposal by proposal id",
                "description": "",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id",
                        "required": true,
                        "type": "string",
                        "format": "string"
                    },
                    {
                        "name": "validator_address",
                        "in": "path",
                        "description": "validator_address",
                        "required": true,
                        "type": "string",
                        "format": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/ApiResponse"
                        }
                    }
                }
            }
        },
        "/validators/{validator_address}": {
            "get": {
                "tags": [
                    "validator"
                ],
                "summary": "get validator info",
                "description": "",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "validator_address",
                        "in": "path",
                        "description": "validator_address",
                        "required": true,
                        "type": "string",
                        "format": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/ApiResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "ApiResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "format": "int32"
                },
                "type": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
    },
...
}